pyproject.toml
[project]
name = "chess-arena-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "autopep8>=2.3.2",
    "flake8>=7.3.0",
    "isort>=7.0.0",
    "mypy>=1.18.2",
    "rich>=14.2.0",
    "python-chess>=1.999",
    "websockets>=13.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["chess_arena_client*"]

[project.scripts]
chess-arena-client = "chess_arena_client.main:main"
chess-arena-client-tester = "chess_arena_client.tester:main"

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-cov>=7.0.0",
]
[tool.pycodestyle]
max-line-length = 120
ignore = ["E251", "E221"]

[tool.isort]
include_trailing_comma = false
use_parentheses = false
line_length = 120

[tool.coverage.run]
omit = ["*/test/*"]

